More sprites, update to bullets
Enemies have health points
Reworked a lot of stuff
The code for the audio of a bullet being fired is written, it's just
that it's commented out. Probably won't uncomment it, because it is
annoying.
Tons and tons of other stuff

To do:
Clean up the code!
When an enemy dies, more should spawn
Kill counter
Some power-ups maybe, or achievements? That's for the distant future
though. Don't know how how much longer I'll keep working on this game
anyways.